Add files via upload
Made progress on making sure the date and times are sent to the python file and that when we recycle the options, it gets updated correctly. 
Things I need to work on:
The logic to get the available spots is not working. Returning an empty directory instead of the correct spots.